Update fmtlib dependency to v10.1.0
I stumpled over the version mismatch when trying to run a binary
which was created with a standalone cmake cross-build for our
Charge Control C platform.

The cross-build binary required fmtlib in version 9 on the target
which was not available.

Reason is, that in meta-everest there is a recipe for v10.1.0 and
it is used when building EVerest in Yocto environments.

Since we are testing our Yocto EVerest builds extensively and
did not encounter any problems so far, I assume that the update
does not cause issues.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>