rust: pin cxxbridge version, update to 1.0.189
When building an EVerest workspace I noticed all Rust modules failed to
link because none of the CXX FFI symbols could be resolved.
The issue appears to be that CMake installs an unbounded (i.e the latest)
version of the `cxxbridge-cmd` tool at configure time, even though it
must match the version of the `cxx` crate used in `everestrs` (cxx#1507).

In release 1.0.189 (the latest as of writing) they started incorporating
the version number in symbol names to detect exactly this scenario, which
broke my build. CMake installed `cxxbridge-cmd` 1.0.189, but the `cxx`
library version was locked to an older release.

This commit fixes that by:

* Updating the `cxx` crate to the latest version. This doesn't solve the
  root cause of the issue, but its nice to have done anyways.
* Locking the version of `cxxbridge-cmd` when installing to match `cxx`,
  and try to update it if the existing binary is outdated. (Re)running
  CMake's configure phase when updating `cxx` now fixes the issue.

In addition I made sure that when installing `cxxbridge-cmd` we always
target the host machine, while performing a fresh build I noticed it
took some cross-compilation settings from the environment and failed.

Signed-off-by: Ivar Scholten <ivar.scholten@protonmail.com>